sp500_dashboard/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
plotly==5.24.1
pandas==2.3.3
jinja2==3.1.4
